requirements.txt
# Please add requirements for installation here
audiomentations
ipython
librosa
notebook
numpy
torchaudio
torchvision
torch_audiomentations
torch
lightning
torchmetrics
torchvision
tensorboard
numpy
matplotlib
scikit-learn
tqdm
Pillow
transformers
datasets
hydra-core
hydra_colorlog
rootuils
wandb
rich
#tf/ perch
apache-beam
clu
flax
ml-collections
optax
tensorflow
tensorflow-io
tensorflow-datasets
pandas
etils
tensorflow-hub
imageio
matplotlib
notebook
chex
numba
